requirements.txt
streamlit>=1.28.0
tensorflow>=2.13.0
numpy>=1.21.0
pillow>=10.0.0
openai>=1.0.0
requests>=2.31.0
gdown>=4.6.0
